Moved State loading utilities from .preparation to .serialization.state
